requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
anyio==4.12.1
asttokens==3.0.1
async-generator==1.10
async-timeout==5.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comtypes==1.4.15
cssselect==1.4.0
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
dotenv==0.9.9
duckduckgo_search==8.1.1
edge-tts==7.2.7
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.3
flashtext==2.7
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
h3==4.4.2
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
importlib_metadata==8.7.1
ipython==8.12.3
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
lxml==6.0.2
lxml_html_clean==0.4.3
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mistune==3.2.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.16.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.2
numpy==2.2.6
opencv-contrib-python==4.13.0.92
outcome==1.3.0.post0
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
pickleshare==0.7.5
pillow==12.1.0
pipreqs==0.5.0
platformdirs==4.5.1
primp==0.15.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycaw==20251023
pycparser==3.0
pydeck==0.9.1
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyperclip==1.11.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-vlc==3.0.21203
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-file==3.0.1
rpds-py==0.30.0
safetensors==0.7.0
scapy==2.7.0
scikit-learn==1.7.2
scipy==1.15.3
selenium==4.40.0
sentence-transformers==5.2.2
sgmllib3k==1.0.0
shellingham==1.5.4
simpleaudio==1.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.5
soupsieve==2.8.3
srt==3.5.3
stack-data==0.6.3
streamlit==1.54.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
threadpoolctl==3.6.0
timezonefinder==8.2.0
tinycss2==1.4.0
tinysegmenter==0.3
tldextract==5.3.1
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.1.0
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
typer-slim==0.21.1
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vosk==0.3.45
watchdog==6.0.0
wcwidth==0.6.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wsproto==1.3.2
yarg==0.1.9
yarl==1.22.0
zipp==3.23.0
